Fixed default type of Image array

+++ pyxel/data_structure/image.py
@@ -24,7 +24,7 @@ class Image(Array):
     Accepted array types: ``np.uint16``, ``np.uint32``, ``np.uint64``
     """
 
-    EXP_TYPE = np.dtype(np.uint)
+    EXP_TYPE = np.dtype(np.uint64)
     TYPE_LIST = (
         np.dtype(np.uint16),
         np.dtype(np.uint32),
